requirements.txt
selenium
webdriver-manage
psutil